+++ network-sdk-sample/app/src/main/AndroidManifest.xml
@@ -11,21 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         android:name=".SampleApplication">
-        <activity
-            android:name="com.braintreepayments.api.BraintreeBrowserSwitchActivity"
-            android:label="@string/title_activity_threedsecure"
-            android:launchMode="singleTask"
-            android:screenOrientation="portrait"
-            tools:replace="android:label">
-            <intent-filter>
-                <action android:name="android.intent.action.VIEW" />
 
-                <category android:name="android.intent.category.DEFAULT" />
-                <category android:name="android.intent.category.BROWSABLE" />
-
-                <data android:scheme="${applicationId}.braintree" />
-            </intent-filter>
-        </activity>
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -33,13 +19,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".booking.AdyenResultActivity">
-
-        </activity>
-
-        <service
-            android:name=".booking.AdyenDropInService"
-            android:permission="android.permission.BIND_JOB_SERVICE"/>
     </application>
 
 </manifest>
